requirements.txt
google-generativeai
python-dotenv
langchain
Flask
twilio
pypdf
faiss-cpu
langchain_google_genai
langchain_community
requests
gunicorn
langchain-google-community
pytesseract
Pillow
fpdf2
beautifulsoup4